tep-1009: Add more tests, update docs and t3db

+++ tests/reset-refs.sh
@@ -8,5 +8,7 @@ find "$test_dir" -name "*.ref.*" | while read ref; do
     refname=$(basename $ref)
     outname=$(echo $refname | sed 's/\(.*\)\.ref\.\(.*\)/out.\1.\2/')
 
-    mv "$dir/$outname" "$dir/$refname"
+    if [ -f $dir/$outname ]; then
+        mv "$dir/$outname" "$dir/$refname"
+    fi
 done
